# src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Desafio Luiza code - Squad C: ENIAC girls",
    "description": "Projeto desenvolvido para o desafio final do Luiza code 3"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "tags": [
    {
      "name": "Usuario",
      "description": "Endpoints relacionados ao recurso de usuario"
    },
    {
      "name": "Loja",
      "description": "Endpoints relacionados ao recurso de loja"
    },
    {
      "name": "Produto",
      "description": "Endpoints relacionados ao recurso de produto"
    },
    {
      "name": "Cliente",
      "description": "Endpoints relacionados ao recurso de cliente"
    },
    {
      "name": "Compra",
      "description": "Endpoints relacionados ao recurso de compra"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization",
      "description": "Insira seu token para garantir acesso aos endpoints"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/lojas/": {
      "get": {
        "tags": [
          "Loja"
        ],
        "description": "Endpoint parra obter uma lista de lojas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Loja encontrada",
            "schema": {
              "$ref": "#/definitions/Loja"
            }
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Loja não encontrada"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/produtos/": {
      "get": {
        "tags": [
          "Produto"
        ],
        "description": "Endpoint parra obter uma lista de produto",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Produto encontrado",
            "schema": {
              "$ref": "#/definitions/Produto"
            }
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/clientes/": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "description": "Endpoint para obter uma lista de clientes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cliente encontrado",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Cliente"
        ],
        "description": "Endpoint para cadastrar um cliente",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "cpf": {
                  "example": "any"
                },
                "endereco": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Cliente"
            },
            "description": "Cliente encontrado"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição"
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/compras/{id}": {
      "get": {
        "tags": [
          "Compra"
        ],
        "description": "Endpoint para obter uma lista de compras à partir do id de um Cliente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Compra encontrada",
            "schema": {
              "$ref": "#/definitions/Compras"
            }
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Compra não encontrada"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/compras/": {
      "post": {
        "tags": [
          "Compra"
        ],
        "description": "Endpoint para incluir uma nova compra",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "ProdutoId": {
                  "example": "any"
                },
                "ClienteId": {
                  "example": "any"
                },
                "valor_total": {
                  "example": "any"
                },
                "pagamento": {
                  "example": "any"
                },
                "LojaId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Compras"
            },
            "description": "Compra encontrada"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Compra não encontrada"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/compras/{cli}/{prod}": {
      "delete": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "cli",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "prod",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/compras/finalizar": {
      "put": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "idCompra": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/compras/entregar": {
      "put": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "idCompra": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/usuarios/": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "description": "Endpoint parra obter uma lista de usuários",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Usuario não encontrado"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Usuario"
        ],
        "description": "Endpoint para cadastrar um novo usuário",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Usuario"
            },
            "description": "Usuario encontrado"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição"
          },
          "404": {
            "description": "Usuario não encontrado"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/usuarios/login": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "description": "Endpoint para gerar um token de usuário",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          },
          "400": {
            "description": "Desculpe, tivemos um problema com a requisição"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Usuario não encontrado"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
    "Produto": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "nome": {
          "type": "string",
          "example": "Notebook Aspire 3"
        },
        "marca": {
          "type": "string",
          "example": "acer"
        },
        "categoria": {
          "type": "string",
          "example": "Informática"
        },
        "preco": {
          "type": "number",
          "example": 5000
        }
      }
    },
    "NovoProduto": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Monitor Pro Gamer"
        },
        "marca": {
          "type": "string",
          "example": "LG"
        },
        "categoria": {
          "type": "string",
          "example": "Informática"
        },
        "preco": {
          "type": "number",
          "example": 1500
        }
      },
      "required": [
        "nome",
        "marca",
        "categoria",
        "preco"
      ]
    }
  }
}